+++ src/query/core/api/index.ts
@@ -1,3 +1,3 @@
 export { Api } from "./Api";
-export { mergeHooks } from "./mergeHooks";
+export { composeHooks } from "./composeHooks";
 export { DEFAULT_RESOURCE_RETENTION_TIME, DEFAULT_COMMAND_RETENTION_TIME } from "./constants";